server/requirements.txt
asgiref==3.5.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
Faker==13.11.1
idna==3.3
nose==1.3.7
psycopg2==2.9.3
PyJWT==2.4.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
register==0.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
